test: reorganization of the 'peer' unit test
Reorganize the 'peer' unit test
and split them into separate binaries.
