+++ tests/unit/CMakeLists.txt
@@ -10,7 +10,7 @@ add_subdirectory(error-test)
 add_subdirectory(flush-test)
 add_subdirectory(info-test)
 add_subdirectory(mr-test)
-add_subdirectory(peer-test)
+add_subdirectory(peer)
 add_subdirectory(private_data-test)
 add_subdirectory(template-test)
 add_subdirectory(utils-test)
